Fixed watchdog crashes (I think);added rolling avg
I'm pretty sure the watchdog crashes were because I was trying
to access the HX711 hardware from two places.

I changed to have the main loop() code access the HX711, and
have the web functions just use the current value.

Changed to read the HX711 frequently, and have the "current value"
be a rolling average.

Lots of tuning oportuities ahead, but this seems to work pretty well.